+++ examples/sveltekit/src/routes/+page.ts
@@ -1,0 +1,13 @@
+import { prefetchQuery } from '$lib/wundergraph';
+import type { PageLoad } from './$types';
+
+export const load: PageLoad = async ({ parent }) => {
+	const { queryClient } = await parent();
+
+	await prefetchQuery(
+		{
+			operationName: 'Dragons',
+		},
+		queryClient
+	);
+};