+++ packages/react-query/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.8.10](https://github.com/wundergraph/wundergraph/compare/@wundergraph/react-query@0.8.9...@wundergraph/react-query@0.8.10) (2023-04-15)
+
+**Note:** Version bump only for package @wundergraph/react-query
+
+## [0.8.9](https://github.com/wundergraph/wundergraph/compare/@wundergraph/react-query@0.8.8...@wundergraph/react-query@0.8.9) (2023-04-14)
+
+**Note:** Version bump only for package @wundergraph/react-query
+
+## [0.8.8](https://github.com/wundergraph/wundergraph/compare/@wundergraph/react-query@0.8.7...@wundergraph/react-query@0.8.8) (2023-04-14)
+
+**Note:** Version bump only for package @wundergraph/react-query
+
+## [0.8.7](https://github.com/wundergraph/wundergraph/compare/@wundergraph/react-query@0.8.6...@wundergraph/react-query@0.8.7) (2023-04-13)
+
+**Note:** Version bump only for package @wundergraph/react-query
+
+## [0.8.6](https://github.com/wundergraph/wundergraph/compare/@wundergraph/react-query@0.8.5...@wundergraph/react-query@0.8.6) (2023-04-12)
+
+**Note:** Version bump only for package @wundergraph/react-query
+
+## [0.8.5](https://github.com/wundergraph/wundergraph/compare/@wundergraph/react-query@0.8.4...@wundergraph/react-query@0.8.5) (2023-04-12)
+
+**Note:** Version bump only for package @wundergraph/react-query
+
 ## [0.8.4](https://github.com/wundergraph/wundergraph/compare/@wundergraph/react-query@0.8.3...@wundergraph/react-query@0.8.4) (2023-04-04)
 
 **Note:** Version bump only for package @wundergraph/react-query