+++ packages/solid-query/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.4.10](https://github.com/wundergraph/wundergraph/compare/@wundergraph/solid-query@0.4.9...@wundergraph/solid-query@0.4.10) (2023-04-15)
+
+**Note:** Version bump only for package @wundergraph/solid-query
+
+## [0.4.9](https://github.com/wundergraph/wundergraph/compare/@wundergraph/solid-query@0.4.8...@wundergraph/solid-query@0.4.9) (2023-04-14)
+
+**Note:** Version bump only for package @wundergraph/solid-query
+
+## [0.4.8](https://github.com/wundergraph/wundergraph/compare/@wundergraph/solid-query@0.4.7...@wundergraph/solid-query@0.4.8) (2023-04-14)
+
+**Note:** Version bump only for package @wundergraph/solid-query
+
+## [0.4.7](https://github.com/wundergraph/wundergraph/compare/@wundergraph/solid-query@0.4.6...@wundergraph/solid-query@0.4.7) (2023-04-13)
+
+**Note:** Version bump only for package @wundergraph/solid-query
+
+## [0.4.6](https://github.com/wundergraph/wundergraph/compare/@wundergraph/solid-query@0.4.5...@wundergraph/solid-query@0.4.6) (2023-04-12)
+
+**Note:** Version bump only for package @wundergraph/solid-query
+
+## [0.4.5](https://github.com/wundergraph/wundergraph/compare/@wundergraph/solid-query@0.4.4...@wundergraph/solid-query@0.4.5) (2023-04-12)
+
+**Note:** Version bump only for package @wundergraph/solid-query
+
 ## [0.4.4](https://github.com/wundergraph/wundergraph/compare/@wundergraph/solid-query@0.4.3...@wundergraph/solid-query@0.4.4) (2023-04-04)
 
 **Note:** Version bump only for package @wundergraph/solid-query